requirements.txt
Django==3.0.5
djangorestframework==3.11.0
django-environ==0.4.5
psycopg2==2.8.4
djangorestframework-simplejwt==4.4.0
